add convert points to cash

+++ Assets/Scripts/Car/Styles/CarStyleObject.cs
@@ -8,6 +8,6 @@ namespace Car
     {
         [field:SerializeField] public CarDetailsEnum Type { get; private set; }
         [field:SerializeField] public CarStyleObjectPrefab Prefab { get; private set; }
-        [field:SerializeField] public float Cost { get; private set; }
+        [field:SerializeField] public int Cost { get; private set; }
     }
 }
