+++ Assets/Scripts/PlayerHub/PlayerData.cs
@@ -5,7 +5,7 @@ namespace PlayerHub
     public class PlayerData
     {
         public string PlayerName { get; set; }
-        public float PlayerCash { get; set; } = 10;
+        public int PlayerCash { get; set; } = 10;
         public int? PlayerCurrentCarId { get; set; }
         public List<int> OpenedCars { get; set; } = new List<int>();
     }
